+++ package.json
@@ -28,8 +28,8 @@
     "typescript": "^5.0.3"
   },
   "scripts": {
-    "build": "ncc build src/main.ts --out dist/main",
-    "watch": "ncc build src/main.ts --out dist/main --watch",
+    "build": "ncc build src/main.js --out dist/main",
+    "watch": "ncc build src/main.js --out dist/main --watch",
     "test": "jest"
   }
 }
