--- run.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-echo "yes | npx tsx ./src/main.ts"
